Update README.md
This Employee Management System is a comprehensive tool built using Core Java and Java Swing to facilitate efficient management of employees and their tasks. Developed as a part of the curriculum at LNCT Group of Colleges, this project showcases the use of object-oriented programming concepts in building a real-world application.